Добавлен сложный файл системы линейных уравнений. Добавлены функции для ситуации когда требуются разные переменные.
