+++ src/App.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import './App.css';
 import { Typography, Container } from '@mui/material';
-import { createExamples as createExamples1 } from '@mathExamples/math-examples1.js'
+import { createExamples as createExamples1 } from '@mathExamples/znacheniye_vyrazheniya.js'
 import { createExamples as createExamples2 } from '@mathExamples/chastnoye_stepeney.js'
 import { createExamples as createExamples3 } from '@mathExamples/proizvedeniye_stepeney.js'
 import { createExamples as createExamples4 } from '@mathExamples/stepen_stepeni.js'
 import { createExamples as createExamples5 } from '@mathExamples/lineynyye_uravneniya.js'
-
+import { createExamples as createExamples6 } from '@mathExamples/sistemy_sposob_slozheniya.js'
 
 import ExampleBlock from "./components/ExampleBlock";
 
@@ -37,6 +37,11 @@ function App() {
         Примеры 5
       </Typography>
       <ExampleBlock createExamples={createExamples5}/>
+
+      <Typography variant="h4" fontWeight="bold" sx={{ fontSize: { lg: '44px', xs: '30px' } }} mb="46px">
+        Примеры 6
+      </Typography>
+      <ExampleBlock createExamples={createExamples6}/>
     </Container>
   )
 }
