+++ src/components/ExamplesCardsGrid.jsx
@@ -6,7 +6,7 @@ const ExamplesCardsGrid = () => {
   const cards = [
     { 
         id: 1,
-        heading: "Простые операции",
+        heading: "Значение выражения",
         description: "Эти упражнения помогут потренировать навык решения простых уравнений"
     },
     { 
@@ -27,6 +27,11 @@ const ExamplesCardsGrid = () => {
     { 
         id: 5,
         heading: "Линейные уравнения",
+        description: "Эти упражнения помогут потренировать навык работать со степенями"
+    },
+    { 
+        id: 6,
+        heading: "Системы уравнений. Способ сложения.",
         description: "Эти упражнения помогут потренировать навык работать со степенями"
     }
