+++ src/math-examples/chastnoye_stepeney.js
@@ -8,9 +8,7 @@ const createExample1 = () => {
 
     const math = `${variable}^{${a}${b}}`
     const answer = `${variable}^{${a+b}}`
-    
-    
-    const condition = `${variable}=${b}`
+    const condition = undefined
   
     return {
       id: uuidv4(),
@@ -28,7 +26,7 @@ const createExample2 = () => {
     
     const math = `${variable}^{${a}+${b}${c}}`
     const answer = `${variable}^{${a+b+c}}`
-    const condition = ``
+    const condition = undefined
 
     return {
       id: uuidv4(),
@@ -45,7 +43,7 @@ const createExample3 = () => {
 
     const math = `${variable}^{${a}}:${variable}^{${b}}`
     const answer = `${variable}^{${a-b}}`;
-    const condition = ``
+    const condition = undefined
 
     return {
       id: uuidv4(),
@@ -62,7 +60,7 @@ const createExample4 = () => {
 
     const math = `${variable}^{${a}}:${variable}^{${b}}`
     const answer = `${variable}^{${a-b}}`
-    const condition = ``
+    const condition = undefined
 
     return {
       id: uuidv4(),
@@ -80,7 +78,7 @@ const createExample5 = () => {
 
   const math = `${variable}^{${b}}:${variable}^{${a}}:${variable}^{${c}}`
   const answer = `${variable}^{${b-a-c}}`
-  const condition = ``
+  const condition = undefined
 
   return {
     id: uuidv4(),
@@ -97,7 +95,7 @@ const createExample6 = () => {
 
   const math = `\\left(\\frac{${variable}}{k}\\right)^{${a}}:\\left(\\frac {${variable}}{k}\\right)^{${b}}`
   const answer = `\\left(\\frac{${variable}}{k}\\right)^{${a-b}}`
-  const condition = ``
+  const condition = undefined
 
   return {
     id: uuidv4(),
@@ -114,7 +112,7 @@ const createExample7 = () => {
 
   const math = `(${variable}m)^{${a}}:(${variable}m)^{${b}}`
   const answer = `(${variable}m)^{${a-b}}`
-  const condition = ``
+  const condition = undefined
 
   return {
     id: uuidv4(),
@@ -131,7 +129,7 @@ const createExample8 = () => {
 
   const math = `(-${variable})^{${a}}:(-${variable})^{${b}}`
   const answer = `(-${variable})^{${a-b}}`
-  const condition = ``
+  const condition = undefined
 
   return {
     id: uuidv4(),
@@ -148,7 +146,7 @@ const createExample9 = () => {
 
   const math = `(${variable}-k)^{${a}}:(${variable}-k)^{${b}}`
   const answer = `(${variable}-k)^{${a-b}}`
-  const condition = ``
+  const condition = undefined
 
   return {
     id: uuidv4(),
@@ -168,7 +166,7 @@ const createExample10 = () => {
 
   const math = `(${variable}${c})^{${a}}:(${variable}${c})^{${b}}`
   const answer = `(${variable}${c})^{${a-b}}`
-  const condition = ``
+  const condition = undefined
 
   return {
     id: uuidv4(),
