+++ src/math-examples/znacheniye_vyrazheniya.js
@@ -2,14 +2,14 @@ import { getRandomInt, generateVariableName, shuffleSingle } from '@utility/math
 import { v4 as uuidv4 } from 'uuid'
 
 const createExample1 = () => {
-    let a = getRandomInt(2, 20);  
-    let b = getRandomInt(2, 20);
+    let a = getRandomInt(-10, 10, 0);  
+    let b = getRandomInt(-30, 30, 0);
     let variable = generateVariableName()
-
-    const math = `${variable}^{${a}+${b}}`
-    const answer = `${variable}^{${a+b}}`
-        
-    const condition = undefined
+  
+    const answer = `${b + a}`;
+  
+    const math = `${variable}${a > 0 ? '+' : ''}${a}`
+    const condition = `${variable}=${b}`
   
     return {
       id: uuidv4(),
@@ -20,14 +20,15 @@ const createExample1 = () => {
   }
 
 const createExample2 = () => {
-    let a = getRandomInt(5, 20);
-    let b = getRandomInt(2, 20);
-    let c = getRandomInt(-6, -2);;
+    let a = getRandomInt(-5, 5, 0);  
+    let b = getRandomInt(-20, 20, 0);
+    let c = getRandomInt(-5, 5, 0);
     let variable = generateVariableName()
-    
-    const math = `${variable}^{${a}+${b}${c}}`
-    const answer = `${variable}^{${a+b+c}}`
-    const condition = undefined
+
+    const answer = `${c*b + a}`;
+
+    const math = `${c}${variable}${a > 0 ? '+' : ''}${a}`
+    const condition = `${variable}=${b}`
 
     return {
       id: uuidv4(),
@@ -38,13 +39,15 @@ const createExample2 = () => {
 }
 
 const createExample3 = () => {
-    let b = getRandomInt(2, 20);
-    let a = getRandomInt(2, 20);
+    let a = getRandomInt(-5, 5, 0);  
+    let b = getRandomInt(-20, 20, 0);
+    let c = getRandomInt(-5, 5, 0);
     let variable = generateVariableName()
 
-    const math = `${variable}^{${a}}${variable}^{${b}}`
-    const answer = `${variable}^{${a+b}}`;
-    const condition = undefined
+    const answer = `${c*b + a}`;
+
+    const math = `${a}${c > 0 ? '+' : ''}${c}${variable}`
+    const condition = `${variable}=${b}`
 
     return {
       id: uuidv4(),
@@ -55,13 +58,14 @@ const createExample3 = () => {
 }
 
 const createExample4 = () => {
-    let b = getRandomInt(2, 20);
-    let a = getRandomInt(2, 20);
+    let a = getRandomInt(-20, 20, 0);
+    let b = getRandomInt(-20, 20, 0);
     let variable = generateVariableName()
 
-    const math = `${variable}^{${a}}${variable}^{${b}}`
-    const answer = `${variable}^{${a+b}}`
-    const condition = undefined
+    const answer = `${a+b}`
+
+    const math = `${variable}+k`
+    const condition = `${variable}=${a}; k=${b}`
 
     return {
       id: uuidv4(),
@@ -72,14 +76,14 @@ const createExample4 = () => {
 }
 
 const createExample5 = () => {
-  let b = getRandomInt(2, 10);
-  let a = getRandomInt(2, 10);
-  let c = getRandomInt(2, 5);
+  let a = getRandomInt(-20, 20, 0);  
+  let b = getRandomInt(-20, -2);
   let variable = generateVariableName()
 
-  const math = `${variable}^{${b}}${variable}^{${a}}${variable}^{${c}}`
-  const answer = `${variable}^{${b+a+c}}`
-  const condition = undefined
+  const answer = `${a-b}`
+
+  const math = `${variable}-w`
+  const condition = `${variable}=${a}; w=${b}`
 
   return {
     id: uuidv4(),
@@ -90,13 +94,14 @@ const createExample5 = () => {
 }
 
 const createExample6 = () => {
-  let b = getRandomInt(2, 20);
-  let a = b*getRandomInt(1, 5);
+  let a = getRandomInt(-20, 20, 0);  
+  let b = getRandomInt(-20, 20, 0);
   let variable = generateVariableName()
 
-  const math = `\\left(\\frac{${variable}}{k}\\right)^{${a}}\\left(\\frac {${variable}}{k}\\right)^{${b}}`
-  const answer = `\\left(\\frac{${variable}}{k}\\right)^{${a+b}}`
-  const condition = ``
+  const answer = `${-1*(a+b)}`
+
+  const math = `-${variable}-k`
+  const condition = `${variable}=${a}; k=${b}`
 
   return {
     id: uuidv4(),
@@ -107,13 +112,14 @@ const createExample6 = () => {
 }
 
 const createExample7 = () => {
-  let b = getRandomInt(2, 20);
-  let a = getRandomInt(2, 20);
+  let a = getRandomInt(-10, 10, 0);  
+  let b = getRandomInt(-30, 30, 0);
   let variable = generateVariableName()
 
-  const math = `(${variable}m)^{${a}}(${variable}m)^{${b}}`
-  const answer = `(${variable}m)^{${a+b}}`
-  const condition = undefined
+  const answer = `${-1*(a+b)}`
+
+  const math = `-(${variable}${a > 0 ? '+' : ''}${a})`
+  const condition = `${variable}=${b};`
 
   return {
     id: uuidv4(),
@@ -124,13 +130,14 @@ const createExample7 = () => {
 }
 
 const createExample8 = () => {
-  let b = getRandomInt(2, 20);
-  let a = getRandomInt(2, 20);
+  let a = getRandomInt(-20, 20, 0);  
+  let b = getRandomInt(-20, 20, 0);
   let variable = generateVariableName()
 
-  const math = `(-${variable})^{${a}}(-${variable})^{${b}}`
-  const answer = `(-${variable})^{${a+b}}`
-  const condition = undefined
+  const answer = `${-1*(a+b)}`
+
+  const math = `-(${variable}+k)`
+  const condition = `${variable}=${a};k=${b}`
 
   return {
     id: uuidv4(),
@@ -141,13 +148,15 @@ const createExample8 = () => {
 }
 
 const createExample9 = () => {
-  let b = getRandomInt(2, 20);
-  let a = getRandomInt(7, 20);
+  const a = getRandomInt(-5, 5, 0);  
+  const b = getRandomInt(-20, 20, 0);
+  const c = getRandomInt(-5, 5, 0);
   let variable = generateVariableName()
 
-  const math = `(${variable}-k)^{${a}}(${variable}-k)^{${b}}`
-  const answer = `(${variable}-k)^{${a+b}}`
-  const condition = undefined
+  const answer = `${c+a+b}`
+
+  const math = `${c}+${variable}${a > 0 ? '+' : ''}${a}`
+  const condition = `${variable}=${b};`
 
   return {
     id: uuidv4(),
@@ -158,16 +167,14 @@ const createExample9 = () => {
 }
 
 const createExample10 = () => {
-  let b = getRandomInt(2, 20);
-  let a = getRandomInt(7, 20);
-  let c = getRandomInt(-30, 30, 0);
-  if(c > 0) c = '+'+c;
-
+  const a = getRandomInt(-10, 10, 0);  
+  const b = getRandomInt(-30, 30, 0);
   let variable = generateVariableName()
 
-  const math = `(${variable}${c})^{${a}}(${variable}${c})^{${b}}`
-  const answer = `(${variable}${c})^{${a+b}}`
-  const condition = undefined
+  const answer = `${b+a}`
+
+  const math = `${variable}${a > 0 ? '+' : ''}${a}`
+  const condition = `${variable}=${b};`
 
   return {
     id: uuidv4(),
@@ -176,7 +183,6 @@ const createExample10 = () => {
     answer: answer
   }
 }
-
 
 const examplesFn = [
   createExample1,
